Change lat&long type in migration file

+++ db/migrate/20160623080418_add_coordinate_to_location.rb
@@ -1,6 +1,6 @@
 class AddCoordinateToLocation < ActiveRecord::Migration
   def change
-    add_column :locations, :lat, :double
-    add_column :locations, :long, :double
+    add_column :locations, :lat, :float
+    add_column :locations, :long, :float
   end
 end
